Add known setuid programs to the `path' variable passed to `activate-configuration'.
This is really a hack since programs not known to `system.nix' will not be in the path
of `activate-configuration', which consequently creates empty `.real' files under
`/var/setuid-wrappers'.  I don't know what the best way to address this is: let admins
write `.real' files by hand?  Add a `setuidPackages' attribute, listing the names of
packages that contain the binaries listed in `setuidPrograms'?


git-svn-id: https://nixos.org/repos/nix/nixos/trunk@11153 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb