Fixing the shutdown script, for the not-that-strange case of having /nix bind mounted from a fs not rootfs.
As niksnut found, if the sources of bind mounted directories are
umounted before remounting them as "ro", a "mount -o remount,ro" on the
bind targets will not leave the filesystem properly umounted, the same
way as a "-o remount,ro" would do on a usual filesystem.

So, the trick of always remounting as 'ro' before attempting an umount seems
to fix the problem.

git-svn-id: https://nixos.org/repos/nix/nixos/trunk@19008 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb