Making the /dev and /dev/shm tmpfs sizes configurable. By default, they take the usual value of "50% of physical RAM".
As /dev/shm can be filled by anyone, and tmpfs does not trigger the OOM killer (and
can hang the machine due to a lack of RAM), I need to configure that down
in order to avoid crashes.

There is still left the /var/run/nscd tmpfs filesystem, also created with 50%
of the RAM, but at least not writeable by anyone. We could find a reasonable
low value for that, or allow configuration.



git-svn-id: https://nixos.org/repos/nix/nixos/trunk@21140 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb